requirements.txt
pwntools
pwndbg @ git+https://github.com/pwndbg/pwndbg.git@dev
